Create db entry on news submission
